Rename files and exports
