+++ src/dashboard/utils/dashboards/index.ts
@@ -1,3 +1,3 @@
 export * from './finance';
 export * from './hr';
-export * from './projectManagement';
+export * from './pm';
